User functionality is fixed. Consultation field was removed. Controller layer has been reorganized. Consultation response and request classes have been created. Consultation getAllConsultations and getConsultById have been completed.

+++ src/main/java/com/maidgroup/maidgroup/dao/ConsultationRepository.java
@@ -15,7 +15,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface ConsultationRepository extends JpaRepository<Consultation, Integer> {
+public interface ConsultationRepository extends JpaRepository<Consultation, Long> {
 
     @Query("select c from Consultation c where c.status = :status")
     public List<Consultation> findByStatus(@Param("status") ConsultationStatus consultationStatus);
